wecom-saas-scaffold/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
PyMySQL==1.1.0
boto3==1.28.57
minio==7.1.17
redis==4.6.0
celery==5.3.4
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
cryptography==41.0.4
pycryptodome==3.19.0
# Placeholder for wecom official SDK; use the official package when available
# wecom-sdk==x.y.z
PyJWT==2.10.1
